requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
Requests==2.32.3
